📻种子数据生成器，初步开发（随机数据功能已完成）
