+++ src/apps/demo/models.py
@@ -70,28 +70,3 @@ class Music(models.Model):
         ordering = ['name']
         verbose_name = 'Music'
         verbose_name_plural = 'Music'
-
-
-import uuid
-
-
-class TestModel(models.Model):
-    char_field = models.CharField(max_length=255)
-    text_field = models.TextField()
-    integer_field = models.IntegerField()
-    float_field = models.FloatField()
-    boolean_field = models.BooleanField()
-    date_field = models.DateField()
-    datetime_field = models.DateTimeField()
-    email_field = models.EmailField()
-    url_field = models.URLField()
-    decimal_field = models.DecimalField(max_digits=10, decimal_places=2)
-    slug_field = models.SlugField(max_length=50)
-    uuid_field = models.UUIDField(default=uuid.uuid4, editable=False, unique=True)
-    binary_field = models.BinaryField(max_length=100)
-    duration_field = models.DurationField()
-    ip_address_field = models.GenericIPAddressField(protocol='both', unpack_ipv4=False)
-    json_field = models.JSONField()
-
-    def __str__(self):
-        return self.char_field
